Fix stack corruption in `mem`.
Problem was read functions were expected to return number of chars read,
but most returned void. These were casted to int, returning random
lengths.

The read syscall handler then wrote the number of bytes returned to the
stack of the userspace program with a number of random bytes, completely
corrupting the stack, leading to a complete system crash.