Segmented Programs (#6)
* Update so that every program runs in a new code segment.

* Now calls program using registers and retf rather than from .data.

* Programs have their own stack now.

* Programs now get stdio correctly at launch.

Argc/v not working yet.

* Simplify interrupt 21h handler.

* Fix open syscall, read syscall up to 128 bytes.

* Make read syscall work for all read sizes.

* Partially fix execa.

* Alter memory map to give kernel more space.

* Re-enable rtc and a20 as more space available now.

* Fix issue where new crt0 was not being used during build.

This wasted a lot of time debugging where code edits were not used in
the program.

* Fix passing argv to crt0.

Still broken passing from crt0 to main for some reason.

* Fix passing args to programs by using memory manager and allocating on
the heap.

* Fix `list_directory` syscall.

* Try to fix read chunking algorithm.

* Bodge the fix to assume `read` is never called over 1024 bytes.

* Fix stack corruption in `mem`.

Problem was read functions were expected to return number of chars read,
but most returned void. These were casted to int, returning random
lengths.

The read syscall handler then wrote the number of bytes returned to the
stack of the userspace program with a number of random bytes, completely
corrupting the stack, leading to a complete system crash.

* Fix highmem/lowmem displaying debug data.

Now bootloader passes the correct information to the kernel.

* Only release when master pushed.

* Add kernel command line, convert sysinfo assembly to C.

* Fix boot on 286 systems.

Due to boot2.bin expanding past a cylinder boundary, int 13h refuses to
read the whole 2nd stage bootloader.

Converted to add a 2nd call to int 13h, which fixes booting on 286s and
below.