Fix malloc corruption finally.
The problem was the `malloc`'ing of `argv[i]` in `proc.c`, which when the
return was over d100h the variable became f52eh for some reason, which
completely destroyed processes with memory corruption and overwriting.

Unfortunately the fix was assining a malloc call to an intermediate
variables which is a workaround, but does appear to fix the problem.